# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "招聘中心系统 API 文档",
        "title": "招聘中心 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/applies": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页获取所有职位申请记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位申请"
                ],
                "summary": "获取申请列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码 (默认值: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量 (默认值: 10)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.JobApply"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的职位申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位申请"
                ],
                "summary": "创建职位申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "申请信息",
                        "name": "apply",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.JobApply"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.JobApply"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/applies/{id}": {
            "get": {
                "description": "获取指定ID的职位申请详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "职位申请"
                ],
                "summary": "获取申请详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "申请ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.JobApply"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errors.ErrorCode": {
            "type": "integer",
            "enum": [
                200,
                400,
                401,
                403,
                404,
                405,
                409,
                429,
                422,
                500,
                503,
                1001,
                1002,
                1003,
                2001,
                2002,
                2003,
                2004,
                2005,
                2006,
                2007,
                2008,
                2009,
                3001,
                3002,
                4001,
                4002,
                4003,
                4004,
                5001,
                5002,
                5003,
                5004,
                5005,
                5006,
                5007,
                5008,
                5009,
                5010,
                5011,
                5012
            ],
            "x-enum-comments": {
                "BadRequest": "请求参数错误",
                "CompanyInactive": "公司未激活",
                "CompanyNotFound": "公司不存在",
                "Conflict": "资源冲突",
                "FileTooLarge": "文件过大",
                "FileTypeNotAllowed": "文件类型不允许",
                "Forbidden": "禁止访问",
                "InternalServerError": "服务器内部错误",
                "InvalidFileFormat": "无效的文件格式",
                "InvalidParams": "无效的请求参数",
                "JobAlreadyApplied": "已经申请过该职位",
                "JobApplicationAlreadyReviewed": "职位申请已被审核",
                "JobApplicationLimitReached": "职位申请次数已达上限",
                "JobApplicationNotFound": "职位申请不存在",
                "JobCreationLimitReached": "职位创建次数已达上限",
                "JobDeleteNotAllowed": "职位删除不允许",
                "JobExpired": "职位已过期",
                "JobNotFound": "职位不存在",
                "JobUpdateNotAllowed": "职位更新不允许",
                "MethodNotAllowed": "方法不允许",
                "NotFound": "资源不存在",
                "PasswordIncorrect": "密码错误",
                "ResumeAccessDenied": "无权限访问简历",
                "ResumeDelete": "简历删除失败",
                "ResumeExists": "简历已存在",
                "ResumeFormat": "简历格式错误",
                "ResumeInvalid": "简历无效",
                "ResumeNotFound": "简历不存在",
                "ResumeParse": "简历解析失败",
                "ResumeTooLarge": "简历文件过大",
                "ResumeUpdate": "简历更新失败",
                "ResumeUpdateStatus": "简历更新状态错误",
                "ResumeUpdateTooFrequent": "简历更新过于频繁",
                "ResumeUpload": "简历上传失败",
                "ServiceUnavailable": "服务不可用",
                "TooManyFiles": "文件数量超过限制",
                "TooManyRequests": "请求过多",
                "Unauthorized": "未授权",
                "UserAlreadyExists": "用户已存在",
                "UserNotFound": "用户不存在"
            },
            "x-enum-descriptions": [
                "",
                "请求参数错误",
                "未授权",
                "禁止访问",
                "资源不存在",
                "方法不允许",
                "资源冲突",
                "请求过多",
                "无效的请求参数",
                "服务器内部错误",
                "服务不可用",
                "用户不存在",
                "用户已存在",
                "密码错误",
                "职位不存在",
                "职位已过期",
                "已经申请过该职位",
                "职位申请次数已达上限",
                "职位创建次数已达上限",
                "职位更新不允许",
                "职位删除不允许",
                "职位申请不存在",
                "职位申请已被审核",
                "公司不存在",
                "公司未激活",
                "文件过大",
                "文件类型不允许",
                "无效的文件格式",
                "文件数量超过限制",
                "简历不存在",
                "简历无效",
                "简历已存在",
                "简历文件过大",
                "简历格式错误",
                "简历上传失败",
                "简历删除失败",
                "无权限访问简历",
                "简历解析失败",
                "简历更新失败",
                "简历更新过于频繁",
                "简历更新状态错误"
            ],
            "x-enum-varnames": [
                "Success",
                "BadRequest",
                "Unauthorized",
                "Forbidden",
                "NotFound",
                "MethodNotAllowed",
                "Conflict",
                "TooManyRequests",
                "InvalidParams",
                "InternalServerError",
                "ServiceUnavailable",
                "UserNotFound",
                "UserAlreadyExists",
                "PasswordIncorrect",
                "JobNotFound",
                "JobExpired",
                "JobAlreadyApplied",
                "JobApplicationLimitReached",
                "JobCreationLimitReached",
                "JobUpdateNotAllowed",
                "JobDeleteNotAllowed",
                "JobApplicationNotFound",
                "JobApplicationAlreadyReviewed",
                "CompanyNotFound",
                "CompanyInactive",
                "FileTooLarge",
                "FileTypeNotAllowed",
                "InvalidFileFormat",
                "TooManyFiles",
                "ResumeNotFound",
                "ResumeInvalid",
                "ResumeExists",
                "ResumeTooLarge",
                "ResumeFormat",
                "ResumeUpload",
                "ResumeDelete",
                "ResumeAccessDenied",
                "ResumeParse",
                "ResumeUpdate",
                "ResumeUpdateTooFrequent",
                "ResumeUpdateStatus"
            ]
        },
...
        "model.JobApply": {
            "type": "object",
            "properties": {
                "applyProgress": {
                    "type": "string"
                },
                "applyTime": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jobId": {
                    "type": "integer"
                },
                "resumeId": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "request.JobApply": {
            "type": "object",
            "properties": {
                "applyProgress": {
                    "description": "改为 applyProgress",
                    "type": "string"
                },
                "applyTime": {
                    "description": "改为 applyTime",
                    "type": "string"
                },
                "createTime": {
                    "description": "改为 createTime",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jobId": {
                    "description": "改为 jobId",
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/request.JobApplyStatus"
                },
                "updateTime": {
                    "description": "改为 updateTime",
                    "type": "string"
                },
                "userId": {
                    "description": "改为 userId",
                    "type": "integer"
                }
            }
        },
        "request.JobApplyStatus": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-comments": {
                "JobApplyStatusAccepted": "已接受",
                "JobApplyStatusInProgress": "进行中",
                "JobApplyStatusPending": "待处理",
                "JobApplyStatusRejected": "已拒绝",
                "JobApplyStatusWithdrawn": "已撤回"
            },
            "x-enum-descriptions": [
                "待处理",
                "进行中",
                "已接受",
                "已拒绝",
                "已撤回"
            ],
            "x-enum-varnames": [
                "JobApplyStatusPending",
                "JobApplyStatusInProgress",
                "JobApplyStatusAccepted",
                "JobApplyStatusRejected",
                "JobApplyStatusWithdrawn"
            ]
        },
...
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorCode"
                        }
                    ]
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorCode"
                        }
                    ]
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        }
    }
}